Add ValueTask extensibility
Adds a new `IValueTaskObject<TResult>` interface and extends `ValueTask<TResult>` to be constructable around such an object in addition to around a `Task<TResult>` and a `TResult`.  This enables `ValueTask<TResult>` to be used with other objects that represent asynchronous operations, and allows for them to be pooled/reused.

Also adds a `IValueTaskObject` and `ValueTask` non-generic equivalent.  A non-generic `ValueTask` has little benefit when it can only be used to represent a `Task`, as `Task.CompletedTask` could always be used to efficiently represented a synchronously, successfully completed operation, but with a non-generic `IValueTaskObject`, a non-generic `ValueTask` has benefits, in that void-returning asynchronous operations can also employ a pool / reusable task-like objects.

There is overhead associated with using `IValueTaskObject`, in particular several interface calls per operation.  Also, `ValueTask`s constructed from `IValueTaskObject`s (both generic and non-generic) are limited in what they can be used for: such `ValueTask`s should only ever be awaited and awaited once; for anything more complicated, the `ValueTask` should be used purely for its `AsTask` method, and then the more complicated operations may be performed on the `Task` or `Task<TResult>` returned from that.  It may be a breaking change for an implementation that was returning `ValueTask<TResult>` around a `Task<TResult>` to instead wrap an `IValueTaskObject<TResult>`, depending on the semantics of the `IValueTaskObject<TResult>`; same for the non-generic counterparts.