And here I split the static method into a deferred instance method call
The instance method approach tends to perform better since it has direct access to the class' fields. Downside is that it means there's more layers, and it's a bit cumbersome to implement. 

This overall strategy only works if the continuation's state doesn't require an additional allocation (e.g. boxing) to make the continuation work. Otherwise, a capturing delegate will achieve essentially the same cost.